backend/user_management/app/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
psycopg[binary]==3.1.12
PyJWT==1.7.1
sqlparse==0.4.4
gunicorn==22.0.0
pillow==10.3.0